requirements.txt
asgiref==3.7.2
Django==3.2.23
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
sqlparse==0.4.4
